added the button icons to the divs in ski body
